code refracorting
